+++ lending/src/sdk/models/operations/downloadexcelreport.ts
@@ -31,7 +31,7 @@ export class DownloadExcelReportResponse extends SpeakeasyBase {
     contentType: string;
 
     /**
-     * Your API request was not properly authorized.
+     * The request made is not valid.
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;