+++ lending/src/sdk/models/operations/listreconciledinvoices.ts
@@ -46,7 +46,7 @@ export class ListReconciledInvoicesResponse extends SpeakeasyBase {
     enhancedInvoicesReport?: shared.EnhancedInvoicesReport;
 
     /**
-     * Your API request was not properly authorized.
+     * Your `query` parameter was not correctly formed
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;