+++ lending/RELEASES.md
@@ -69,3 +69,13 @@ Based on:
 - [typescript v3.3.0] lending
 ### Releases
 - [NPM v3.3.0] https://www.npmjs.com/package/@codat/lending/v/3.3.0 - lending
+
+## 2023-11-06 01:29:01
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.112.1 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.4.0] lending
+### Releases
+- [NPM v3.4.0] https://www.npmjs.com/package/@codat/lending/v/3.4.0 - lending